chore: add item request service
