+++ apps/riven/lib/services/item-request/utilities/persist-requested-movie.ts
@@ -5,23 +5,16 @@ import { ItemRequestCreateError } from "@repo/util-plugin-sdk/schemas/events/ite
 import { ValidationError, validateOrReject } from "class-validator";
 import z from "zod";
 
-import { database } from "../../../../database/database.ts";
-import { logger } from "../../../../utilities/logger/logger.ts";
-import { RequestType } from "../request-content-services.schema.ts";
+import { RequestType } from "../../../message-queue/flows/request-content-services/request-content-services.schema.ts";
 
+import type { EntityManager } from "@mikro-orm/core";
 import type { ContentServiceRequestedResponse } from "@repo/util-plugin-sdk/schemas/events/content-service-requested.event";
 
 export async function persistRequestedMovie(
+  em: EntityManager,
   item: ContentServiceRequestedResponse["movies"][number],
 ) {
-  const externalIds = [
-    item.imdbId ? `IMDB: ${item.imdbId}` : null,
-    item.tmdbId ? `TMDB: ${item.tmdbId}` : null,
-  ].filter(Boolean);
-
-  logger.silly(`Processing requested movie: ${externalIds.join(", ")}`);
-
-  const existingItem = await database.itemRequest.findOne({
+  const existingItem = await em.findOne(ItemRequest, {
     $or: [
       ...(item.imdbId ? [{ imdbId: item.imdbId }] : []),
       ...(item.tmdbId ? [{ tmdbId: item.tmdbId }] : []),
@@ -36,8 +29,6 @@ export async function persistRequestedMovie(
     });
   }
 
-  const em = database.em.fork();
-
   const itemRequest = em.create(ItemRequest, {
     state: "requested",
     requestedBy: item.requestedBy ?? null,
@@ -49,10 +40,6 @@ export async function persistRequestedMovie(
 
   try {
     await validateOrReject(itemRequest);
-
-    await em.flush();
-
-    await em.refreshOrFail(itemRequest);
 
     return {
       requestType: RequestType.enum.create,
